+++ client/src/ShiningText.css
