+++ client/src/ThanosSnap.css
